Mega Code Cleanup, the Finale
