+++ Mr_Steeltastic/commands/armCommands.py
@@ -1,8 +1,7 @@
 import commands2
-import constants
 import ctre
-import wpilib
 from subsystems.arm import Arm, ArmMotor
+import wpilib
 
 """
 COMMANDS
@@ -23,7 +22,7 @@ class SetPosition(commands2.CommandBase):
         self.done = False
         self.timer.start()
 
-        self.motor.toPos(self.pos)
+        self.motor.moveToPos(self.pos)
 
         wpilib.SmartDashboard.putBoolean(f"Moving {self.motor.name}?", True)
         wpilib.SmartDashboard.putString("Current Command", f"Position{self.motor.name}")
@@ -31,7 +30,7 @@ class SetPosition(commands2.CommandBase):
     def execute(self):
         if not self.motor.isMotorPosInRange(self.pos):
             if self.timer.get() >= 2:
-                self.motor.toPos(self.pos)
+                self.motor.moveToPos(self.pos)
                 self.timer.reset()
         else:
             self.done = True
@@ -41,7 +40,7 @@ class SetPosition(commands2.CommandBase):
         wpilib.SmartDashboard.putString("Current Command", "None")
         self.timer.stop()
         self.timer.reset()
-        self.motor.toPos(self.motor.motor.getSelectedSensorPosition())
+        self.motor.moveToPos(self.motor.motor.getSelectedSensorPosition())
         
     def isFinished(self):
         """
@@ -70,7 +69,7 @@ class InvertSafelyThenSetPos(commands2.CommandBase):
 
         self.has_inverted = False
 
-        self.motor.moveToPos(pos=0)
+        self.motor.moveToPos(0)
 
     def execute(self) -> None:
         if self.motor.motor.getInverted() == self.invert and not self.has_inverted:
@@ -79,113 +78,18 @@ class InvertSafelyThenSetPos(commands2.CommandBase):
 
         if self.motor.isMotorPosInRange(0) and not self.has_inverted:
             self.motor.motor.setInverted(self.invert)
-            self.motor.moveToPos(pos=-self.pos)
+            self.motor.moveToPos(-self.pos)
             self.has_inverted = True
 
         if self.motor.isMotorPosInRange(self.pos) and self.has_inverted:
             self.done = True
 
     def end(self, interrupted) -> None:
-        self.motor.toPos(-self.pos)
+        self.motor.moveToPos(-self.pos)
 
     def isFinished(self):
         return self.done
-
-class SetPositionCubePickup(commands2.CommandBase):
-    """
-    Moves the base motor into the given position.
-    """
-    def __init__(self, arm: Arm) -> None:
-        
-        super().__init__()
-        
-        self.arm = arm
-        self.addRequirements([self.arm])
-
-        wpilib.SmartDashboard.putBoolean("Moving Base?", False)
-        wpilib.SmartDashboard.putBoolean("Moving Mid?", False)
-        wpilib.SmartDashboard.putBoolean("Moving Top?", False)
-        self.baseAngle = 44500
-        self.midAngle = -8750
-        self.topAngle = 8821
-        self.arm.baseMotor.motor.setNeutralMode(ctre.NeutralMode.Brake)
-        self.arm.midMotor.motor.setNeutralMode(ctre.NeutralMode.Brake)
-        self.arm.topMotor.motor.setNeutralMode(ctre.NeutralMode.Brake)
-        self.timer = wpilib.Timer()
-        self.baseMotorDone = False
-        self.midMotorDone = False
-        self.topMotorDone = False
-        
-    def initialize(self) -> None:
-        self.baseMotorDone = False
-        self.midMotorDone = False
-        self.topMotorDone = False
-        self.baseMidMotorSent = False
-        self.arm.motorToPos(self.arm.topMotor, self.topAngle/constants.TOPRATIO)
-        self.timer.start()
-        wpilib.SmartDashboard.putBoolean("Moving Base?", False)
-        wpilib.SmartDashboard.putBoolean("Moving Mid?", False)
-        wpilib.SmartDashboard.putBoolean("Moving Top?", True)
-        wpilib.SmartDashboard.putString("Current Command", "CubePickup")
-
-    def execute(self):
-        wpilib.SmartDashboard.putNumber("Timer", self.timer.get())
-
-        if not self.arm.topMotor.isMotorPosInRange(self.topAngle/constants.TOPRATIO):
-            if self.timer.get() >= 3:
-                self.arm.motorToPos(self.arm.topMotor, self.topAngle/constants.TOPRATIO)
-        else:
-            self.topMotorDone = True
-
-        if self.topMotorDone and not self.baseMidMotorSent:
-            self.arm.motorToPos(self.arm.baseMotor, self.baseAngle/constants.BASERATIO)
-            self.arm.motorToPos(self.arm.midMotor, self.midAngle/constants.MIDDLERATIO)
-            self.baseMidMotorSent = True
-
-        if self.arm.midMotor.isMotorPosInRange(self.midAngle/constants.MIDDLERATIO):
-            self.midMotorDone = True
-        
-        if self.arm.baseMotor.isMotorPosInRange(self.baseAngle/constants.BASERATIO):
-            self.baseMotorDone = True
-
     
-    def end(self, interrupted):
-        wpilib.SmartDashboard.putBoolean("Moving Arm?", False)
-        wpilib.SmartDashboard.putBoolean("Moving Base?", False)
-        wpilib.SmartDashboard.putBoolean("Moving Mid?", False)
-        wpilib.SmartDashboard.putBoolean("Moving Top?", False)
-        wpilib.SmartDashboard.putString("Current Command", "None")
-        self.timer.stop()
-        self.timer.reset()
-        wpilib.SmartDashboard.putNumber("Base Time", 0)
-        self.arm.holdAtPos()
-        
-    def isFinished(self):
-        """
-        Return whether or not the command is finished.
-        """
-        return self.baseMotorDone and self.midMotorDone and self.topMotorDone
-    
-"""
-COMMAND GROUPS
-"""
-
-class SetPositionBaseMid(commands2.ParallelCommandGroup):
-    def __init__(self, arm: Arm, basePos, midPos):
-        super().__init__()
-        self.addCommands(
-            SetPosition(arm.baseMotor, basePos),
-            SetPosition(arm.midMotor, midPos)
-            )
-        
-class SetPositionMidTop(commands2.ParallelCommandGroup):
-    def __init__(self, arm: Arm, midPos, topPos):
-        super().__init__()
-        self.addCommands(
-            SetPosition(arm.midMotor, midPos), 
-            SetPosition(arm.topMotor, topPos)
-        )
-
 class SetPositionAll(commands2.ParallelCommandGroup):
     def __init__(self, arm: Arm, basePos, midPos, topPos):
         super().__init__()
@@ -195,6 +99,43 @@ class SetPositionAll(commands2.ParallelCommandGroup):
             SetPosition(arm.topMotor, topPos),
             SetPosition(arm.grabberMotor, 0)
         )
+
+class ToggleArmCoast(commands2.CommandBase):
+    """
+    Sets all arm motors to coast (this lets us move the arm when disabled)
+    """
+    def __init__(self, arm: Arm) -> None:
+        super().__init__()
+
+        self.arm = arm
+        self.addRequirements([self.arm])
+        self.isInCoast = False
+        self.done = False
+
+    def initialize(self) -> None:
+        self.done = False
+
+    def execute(self):
+
+        if self.isInCoast:
+            self.arm.baseMotor.motor.setNeutralMode(ctre.NeutralMode.Brake)
+            self.arm.midMotor.motor.setNeutralMode(ctre.NeutralMode.Brake)
+            self.arm.topMotor.motor.setNeutralMode(ctre.NeutralMode.Brake)
+            self.arm.grabberMotor.motor.setNeutralMode(ctre.NeutralMode.Brake)
+            wpilib.SmartDashboard.putBoolean("In Coast", False)
+
+        else:
+            self.arm.baseMotor.motor.setNeutralMode(ctre.NeutralMode.Coast)
+            self.arm.midMotor.motor.setNeutralMode(ctre.NeutralMode.Coast)
+            self.arm.topMotor.motor.setNeutralMode(ctre.NeutralMode.Coast)
+            self.arm.grabberMotor.motor.setNeutralMode(ctre.NeutralMode.Coast)
+            wpilib.SmartDashboard.putBoolean("In Coast", True)
+
+        self.isInCoast = not self.isInCoast
+        self.done = True
+
+    def isFinished(self) -> bool:
+        return self.done
 
 """
 CONTROLLER-CALLED COMMAND GROUPS (commands groups that run from controller inputs)
@@ -232,5 +173,6 @@ class MoveCubePickup(commands2.SequentialCommandGroup):
         super().__init__()
         self.addCommands(
             SetPositionAll(arm, 0, -30340, 0),
-            SetPositionCubePickup(arm)
+            SetPosition(arm.topMotor, 8821),
+            SetPosition(arm.midMotor, -8750).alongWith(SetPosition(arm.baseMotor, 44500))
         )
