+++ Mr_Steeltastic/constants.py
@@ -52,13 +52,10 @@ ARMBASEF = 0.23
 
 ARMMIDP = 0.8
 ARMMIDD = 0
-#ARMMIDF = 0.05244
 ARMMIDF = 0.05
 
-# ARMTOPP = 1.5
 ARMTOPP = 0.85
 ARMTOPD = 0
-#ARMTOPF = 0.27962
 ARMTOPF = 1.25029
 
 ARMGRABBERP = 0.29
@@ -74,14 +71,12 @@ ARMBASECRUISEVEL = 9000
 ARMMIDCRUISEVEL = 4000
 ARMTOPCRUISEVEL = 2000
 ARMGRABBERCRUISEVEL = 200
-#ARMWRISTCRUISEVEL = 10
 
 # Motion Acceleration
 ARMBASEACCEL = 9000
 ARMMIDACCEL = 4000
 ARMTOPACCEL = 2000
 ARMGRABBERACCEL = 180
-#ARMWRISTACCEL = 100
 
 # Hold Percentages
 ARMBASEHOLDPERCENT = -0.135
@@ -90,7 +85,6 @@ ARMTOPHOLDPERCENT = 0.125
 ARMGRABBERHOLDPERCENT = 0
 
 # Solenoid on Grabber
-
 SOLENOIDMODULE = 0
 SOLENOIDMODULETYPE = wpilib.PneumaticsModuleType.CTREPCM
 GRABBERSOLENOIDIN = 0
