--- Mr_Steeltastic/holdPercentages.txt
@@ -1,25 +1,0 @@
------------------
-HOLD PERCENTAGES FOR MR.STEELTASTIC (written by Caden D, so blame me for the amazing format of this file)
------------------
-
-NO CONE:
-
-GWABBA: -0.1075 (had issues when testing top motor, so like, be careful ig. we have it set to -0.1 rn in the code but it also had slight issues BUT it worked slightly better woo)
-
-TOP: 0.125
-
-TOP with cone: 0.145
-
-MID: -0.105
-
-MID with cone: -0.113
-
-BASE: -0.135
-
-BASE with cone: -0.135
-
-
-
-
-
-
