--- Mr_Steeltastic/subsystems/vision.py
@@ -1,4 +1,0 @@
-import wpilib
-
-def cameraLaunch():
-    wpilib.CameraServer.launch()
